Ford Fulkerson with residual edges
The Ford Fulkerson algorithm is a greedy algorithm that determines the maximum flow possible in a flow network.

As long as there is a path from the source node (defaulted to node 1) to the sink (defaulted to node n) with available capacity on all edges in the path, we send flow along one of the paths. We then find another path and repeat the process until maximum flow is achieved.